requirements.txt
appdirs==1.4.1
Django==1.8
django-ckeditor==5.2.1
feedparser==5.2.1
olefile==0.44
packaging==16.8
Pillow==4.0.0
psycopg2==2.6.2
pyparsing==2.1.10
reportlab==3.3.0
six==1.10.0

